Initial support for Proxmox Virtual Environment
Add a machinery module allowing to access Proxmox-based virtual machines via
its REST API.

Based on #1831. Does not, however, rely on the Proxmox VM ID to find the VM and
instead uses the configured label, avoiding the need to extend the Cuckoo
database schema with this machinery specific information.

Adds the proxmoxer python module for REST API access as a
dependency.